make left-panel related widgets into widget classes instead of returning widgets from functions
