+++ lib/routes/world/left_panel/left_panel_add_course_subpage.dart
@@ -1,9 +1,13 @@
 import 'package:flutter/material.dart';
 
 import 'package:fluffychat/features/course_plans/new_course_page.dart';
+import 'package:fluffychat/features/navigation/panel_token.dart';
+import 'package:fluffychat/l10n/l10n.dart';
 import 'package:fluffychat/routes/courses/find_course_page.dart';
 import 'package:fluffychat/routes/courses/private/course_code_page.dart';
-import 'package:fluffychat/routes/world/courses_panel_view.dart';
+import 'package:fluffychat/routes/world/left_panel/left_panel_close_button.dart';
+import 'package:fluffychat/routes/world/left_panel/left_panel_courses_list_view.dart';
+import 'package:fluffychat/routes/world/panel_header.dart';
 
 /// The body of the left-column **add-course panel** (world_v2): the add-course
 /// wizard, hosted as a URL-token panel instead of the retired route-driven
@@ -13,19 +17,23 @@ import 'package:fluffychat/routes/world/courses_panel_view.dart';
 /// code). Each hosted page carries its own header/close; the deeper steps
 /// (`/courses/own/:courseid` …) stay route-driven detail. See
 /// `routing.instructions.md`.
-class AddCoursePanel extends StatelessWidget {
-  /// The wizard step from the token param: null (the hub), `own`, `browse`, or
-  /// `private`.
-  final String? subPath;
-
-  /// The current URL, so the plan list can read its `lang`/`showAll` filters.
+class LeftPanelAddCourseSubpage extends StatelessWidget {
+  final PanelToken token;
   final Uri currentUri;
+  final bool foldedOver;
+  final bool isColumnMode;
 
-  const AddCoursePanel({super.key, this.subPath, required this.currentUri});
+  const LeftPanelAddCourseSubpage({
+    super.key,
+    required this.token,
+    required this.currentUri,
+    required this.foldedOver,
+    required this.isColumnMode,
+  });
 
   @override
   Widget build(BuildContext context) {
-    switch (subPath) {
+    switch (token.param) {
       case 'browse':
         return const FindCoursePage();
       case 'private':
@@ -37,10 +45,20 @@ class AddCoursePanel extends StatelessWidget {
           showAll: currentUri.queryParameters['showAll'] == 'true',
         );
       default:
-        // A bare `addcourse` token is the Courses panel. It is normally
-        // rendered by the panel host with a "Courses" header + close; this is a
-        // headerless fallback for any direct use. See CoursesPanelView.
-        return const CoursesPanelView();
+        return Column(
+          children: [
+            PanelHeader(
+              leading: LeftPanelCloseButton(
+                token: token,
+                currentUri: currentUri,
+                foldedOver: foldedOver,
+                isColumnMode: isColumnMode,
+              ),
+              title: L10n.of(context).courses,
+            ),
+            Expanded(child: LeftPanelCoursesListView()),
+          ],
+        );
     }
   }
 }
